Spruce up the repo
Good looks mean everything
